requirements.txt
Django>=3.0,<3.1
psycopg2>=2.7,<3.0
gunicorn>=19.0,<20.0
graphene-django>=2.0
django-filter>=2
graphene>=2.1,<3
graphql-core>=2.1,<3
django-graphql-jwt>=0.3.1,<0.4